dioxus

Module events

Modules

keyboard_types

Contains types to define keyboard related events.

Structs

AnimationData

ClipboardData

CompositionData

DOMRect

DragData

The DragEvent interface is a DOM event that represents a drag and drop interaction. The user initiates a drag by placing a pointer device (such as a mouse) on the touch surface and then dragging the pointer to a new location (such as another DOM element). Applications are free to interpret a drag and drop interaction in an application-specific way.

FocusData

FormData

FormValue

A form value that may either be a list of values or a single value

ImageData

KeyboardData

MediaData

Modifiers

Pressed modifier keys.

MountedData

An Element that has been rendered and allows reading and modifying information about it.

MouseData

A synthetic event that wraps a web-style MouseEvent Data associated with a mouse event

PlatformEventData

A platform specific event.

PointerData

ResizeData

ScrollData

SelectionData

SerializedAnimationData

A serialized version of AnimationData

SerializedClipboardData

A serialized version of ClipboardData

SerializedCompositionData

A serialized version of CompositionData

SerializedDragData

A serialized version of DragData

SerializedFocusData

A serialized version of FocusData

SerializedFormData

A serialized form data object

SerializedImageData

A serialized version of ImageData

SerializedKeyboardData

A serialized version of KeyboardData

SerializedMediaData

A serialized version of MediaData

SerializedMouseData

A serialized version of MouseData

SerializedPointInteraction

SerializedPointerData

A serialized version of PointerData

SerializedResizeData

A serialized version of ResizeData

SerializedScrollData

A serialized version of ScrollData

SerializedSelectionData

A serialized version of SelectionData

SerializedToggleData

A serialized version of ToggleData

SerializedTouchData

A serialized version of TouchData

SerializedTransitionData

A serialized version of TransitionData

SerializedVisibleData

A serialized version of VisibleData

SerializedWheelData

A serialized version of WheelData

ToggleData

TouchData

TouchPoint

TransitionData

VisibleData

WheelData

Data associated with a WheelEvent

Enums

Code

Code is the physical position of a key.

Key

Key represents the meaning of a keypress.

KeyCode

Location

The location attribute contains an indication of the logical location of the key on the device.

MountedError

The error type for the MountedData

ResizeError

The error type for the MountedData

ScrollBehavior

The way that scrolling should be performed

VisibleError

The error type for the VisibleData

Traits

AExtension

AbbrExtension

AddressExtension

AnimateExtension

AnimateMotionExtension

AnimateTransformExtension

AnnotationExtension

AnnotationXmlExtension

AreaExtension

ArticleExtension

AsideExtension

AudioExtension

BExtension

BaseExtension

BdiExtension

BdoExtension

BlockquoteExtension

BodyExtension

BrExtension

ButtonExtension

CanvasExtension

CaptionExtension

CircleExtension

CiteExtension

ClipPathExtension

CodeExtension

ColExtension

ColgroupExtension

DataExtension

DatalistExtension

DdExtension

DefsExtension

DelExtension

DescExtension

DetailsExtension

DfnExtension

DialogExtension

DiscardExtension

DivExtension

DlExtension

DtExtension

EllipseExtension

EmExtension

EmbedExtension

FeBlendExtension

FeColorMatrixExtension

FeComponentTransferExtension

FeCompositeExtension

FeConvolveMatrixExtension

FeDiffuseLightingExtension

FeDisplacementMapExtension

FeDistantLightExtension

FeDropShadowExtension

FeFloodExtension

FeFuncAExtension

FeFuncBExtension

FeFuncGExtension

FeFuncRExtension

FeGaussianBlurExtension

FeImageExtension

FeMergeExtension

FeMergeNodeExtension

FeMorphologyExtension

FeOffsetExtension

FePointLightExtension

FeSpecularLightingExtension

FeSpotLightExtension

FeTileExtension

FeTurbulenceExtension

FieldsetExtension

FigcaptionExtension

FigureExtension

FilterExtension

FooterExtension

ForeignObjectExtension

FormExtension

GExtension

GlobalAttributesExtension

H1Extension

H2Extension

H3Extension

H4Extension

H5Extension

H6Extension

HasAnimationData

A trait for any object that has the data for an animation event

HasClipboardData

HasCompositionData

A trait for any object that has the data for a composition event

HasDragData

A trait for any object that has the data for a drag event

HasFocusData

HasFormData

An object that has all the data for a form event

HasImageData

A trait for any object that has the data for an image event

HasKeyboardData

HasMediaData

HasMouseData

A trait for any object that has the data for a mouse event

HasPointerData

A trait for any object that has the data for a pointer event

HasResizeData

HasScrollData

HasSelectionData

HasToggleData

HasTouchData

HasTouchPointData

A trait for touch point data

HasTransitionData

HasVisibleData

HasWheelData

HatchExtension

HatchpathExtension

HeadExtension

HeaderExtension

HgroupExtension

HrExtension

HtmlEventConverter

A converter between a platform specific event and a general event. All code in a renderer that has a large binary size should be placed in this trait. Each of these functions should be snipped in high levels of optimization.

IExtension

IframeExtension

ImageExtension

ImgExtension

InputExtension

InsExtension

InteractionElementOffset

A interaction that contains data about the location of the event.

InteractionLocation

A interaction that contains data about the location of the event.

KbdExtension

LabelExtension

LegendExtension

LiExtension

LineExtension

LinearGradientExtension

LinkExtension

MainExtension

MapExtension

MarkExtension

MarkerExtension

MaskExtension

MathExtension

MenuExtension

MerrorExtension

MetaExtension

MetadataExtension

MeterExtension

MfracExtension

MiExtension

MmultiscriptsExtension

MnExtension

MoExtension

ModifiersInteraction

A interaction that contains data about the current state of the keyboard modifiers.

MoverExtension

MpaddedExtension

MpathExtension

MphantomExtension

MprescriptsExtension

MrootExtension

MrowExtension

MsExtension

MspaceExtension

MsqrtExtension

MstyleExtension

MsubExtension

MsubsupExtension

MsupExtension

MtableExtension

MtdExtension

MtextExtension

MtrExtension

MunderExtension

MunderoverExtension

NavExtension

NoscriptExtension

ObjectExtension

OlExtension

OptgroupExtension

OptionExtension

OutputExtension

PExtension

ParamExtension

PathExtension

PatternExtension

PictureExtension

PointerInteraction

A interaction that contains data about the pointer button(s) that triggered the event.

PolygonExtension

PolylineExtension

PreExtension

ProgressExtension

QExtension

RadialGradientExtension

RectExtension

RenderedElementBacking

An Element that has been rendered and allows reading and modifying information about it.

RpExtension

RtExtension

RubyExtension

SExtension

SampExtension

ScriptExtension

SectionExtension

SelectExtension

SemanticsExtension

SetExtension

SlotExtension

SmallExtension

SourceExtension

SpanExtension

StopExtension

StrongExtension

StyleExtension

SubExtension

SummaryExtension

SupExtension

SvgAttributesExtension

SvgExtension

SwitchExtension

SymbolExtension

TableExtension

TbodyExtension

TdExtension

TemplateExtension

TextExtension

TextPathExtension

TextareaExtension

TfootExtension

ThExtension

TheadExtension

TimeExtension

TitleExtension

TrExtension

TrackExtension

TspanExtension

UExtension

UlExtension

UseExtension

VarExtension

VideoExtension

ViewExtension

WbrExtension

Functions

onabort

abort

onanimationend

onanimationend

onanimationiteration

onanimationiteration

onanimationstart

onanimationstart

onblur

onblur

oncanplay

canplay

oncanplaythrough

canplaythrough

onchange

onchange

onclick

Execute a callback when a button is clicked.

oncompositionend

oncompositionend

oncompositionstart

oncompositionstart

oncompositionupdate

oncompositionupdate

oncontextmenu

oncontextmenu

oncopy

oncopy

oncut

oncut

ondblclickDeprecated

General Event Handler Information

ondoubleclick

General Event Handler Information

ondrag

ondrag

ondragend

ondragend

ondragenter

ondragenter

ondragexit

ondragexit

ondragleave

ondragleave

ondragover

ondragover

ondragstart

ondragstart

ondrop

ondrop

ondurationchange

durationchange

onemptied

emptied

onencrypted

encrypted

onended

ended

onerror

onerror

onfocus

onfocus

onfocusin

General Event Handler Information

onfocusout

General Event Handler Information

ongotpointercapture

gotpointercapture

oninput

The oninput event is fired when the value of a <input>, <select>, or <textarea> element is changed.

oninvalid

oninvalid

onkeydown

onkeydown

onkeypress

onkeypress

onkeyup

onkeyup

onload

onload

onloadeddata

loadeddata

onloadedmetadata

loadedmetadata

onloadstart

loadstart

onlostpointercapture

lostpointercapture

onmounted

The onmounted event is fired when the element is first added to the DOM. This event gives you a MountedData object and lets you interact with the raw DOM element.

onmousedown

onmousedown

onmouseenter

onmouseenter

onmouseleave

onmouseleave

onmousemove

onmousemove

onmouseout

onmouseout

onmouseover

onmouseover

onmouseup

onmouseup

onpaste

onpaste

onpause

pause

onplay

play

onplaying

playing

onpointercancel

pointercancel

onpointerdown

pointerdown

onpointerenter

pointerenter

onpointerleave

pointerleave

onpointermove

pointermove

onpointerout

pointerout

onpointerover

pointerover

onpointerup

pointerup

onprogress

progress

onratechange

ratechange

onreset

onreset

onresize

onresize

onscroll

onscroll

onseeked

seeked

onseeking

seeking

onselect

select

onselectionchange

selectionchange

onselectstart

selectstart

onstalled

stalled

onsubmit

onsubmit

onsuspend

suspend

ontimeupdate

timeupdate

ontoggle

ontoggle

ontouchcancel

touchcancel

ontouchend

touchend

ontouchmove

touchmove

ontouchstart

touchstart

ontransitionend

transitionend

onvisible

onvisible

onvolumechange

volumechange

onwaiting

waiting

onwheel

Called when the mouse wheel is rotated over an element.

set_event_converter

Type Aliases

AnimationEvent

ClipboardEvent

CompositionEvent

DragEvent

FocusEvent

FormEvent

ImageEvent

KeyboardEvent

MediaEvent

MountedEvent

MountedResult

The MountedResult type for the MountedData

MouseEvent

PointerEvent

A synthetic event that wraps a web-style PointerEvent

ResizeEvent

ResizeResult

The ResizeResult type for the ResizeData

ScrollEvent

SelectionEvent

ToggleEvent

TouchEvent

TransitionEvent

VisibleEvent

VisibleResult

The VisibleResult type for the VisibleData

WheelEvent

A synthetic event that wraps a web-style WheelEvent