Trait dioxus::events::HasTouchPointData

pub trait HasTouchPointData: InteractionLocation + Any {
    // Required methods
    fn identifier(&self) -> i32;
    fn force(&self) -> f64;
    fn radius(&self) -> Point2D<f64, ScreenSpace>;
    fn rotation(&self) -> f64;
    fn as_any(&self) -> &(dyn Any + 'static);
}

A trait for touch point data

Required Methods

fn identifier(&self) -> i32

A unique identifier for this touch point that will be the same for the duration of the touch

fn force(&self) -> f64

the pressure of the touch

fn radius(&self) -> Point2D<f64, ScreenSpace>

the radius of the touch

fn rotation(&self) -> f64

the rotation of the touch in degrees between 0 and 90

fn as_any(&self) -> &(dyn Any + 'static)

return self as Any