Type Alias dioxus::events::FormEvent

pub type FormEvent = Event<FormData>;

Aliased Type

struct FormEvent {
    pub data: Rc<FormData, Global>,
    /* private fields */
}

Fields
data: Rc<FormData, Global>

The data associated with this event

Implementations

impl<T> Event<T>

pub fn cancel_bubble(&self)

👎Deprecated: use stop_propagation instead

Prevent this event from continuing to bubble up the tree to parent elements.

Example

rsx! {
    button {
        onclick: move |evt: Event<MouseData>| {
            evt.cancel_bubble();
        }
    }
}

pub fn stop_propagation(&self)

Prevent this event from continuing to bubble up the tree to parent elements.

Example

rsx! {
    button {
        onclick: move |evt: Event<MouseData>| {
            evt.stop_propagation();
        }
    }
}

pub fn inner(&self) -> &Rc<T, Global>

Get a reference to the inner data from this event

rsx! {
    button {
        onclick: move |evt: Event<MouseData>| {
            let data = evt.inner.clone();
            cx.spawn(async move {
                println!("{:?}", data);
            });
        }
    }
}