dioxus

Function Ok

pub fn Ok<T>(value: T) -> Result<T, CapturedError>

A helper function for an Ok result that can be either a boxed error or a value This is useful to avoid having to use Ok<T, CapturedError> everywhere