Function dioxus::prelude::GoForwardButton

pub fn GoForwardButton(props: HistoryButtonProps) -> Option<VNode>

A button to go forward through the navigation history. Similar to a browsers forward button.

Only works as descendant of a [Link] component, otherwise it will be inactive.

The button will disable itself if it is known that no later history is available.

Panic

When the GoForwardButton is not nested within a [Link] component hook, but only in debug builds.

Example

#[derive(Clone, Routable)]
enum Route {
    #[route("/")]
    Index {},
}

#[component]
fn App() -> Element {
    rsx! {
        Router::<Route> {}
    }
}

#[component]
fn Index() -> Element {
    rsx! {
        GoForwardButton {
            "go forward"
        }
    }
}