dioxus::prelude

Struct CopyValue

pub struct CopyValue<T, S = UnsyncStorage>
where
    T: 'static,
    S: Storage<T>,
{ /* private fields */ }

CopyValue is a wrapper around a value to make the value mutable and Copy.

It is internally backed by generational_box::GenerationalBox.

Implementations

impl<T> CopyValue<T>
where T: 'static,

pub fn new(value: T) -> CopyValue<T>

Create a new CopyValue. The value will be stored in the current component.

Once the component this value is created in is dropped, the value will be dropped.

pub fn new_in_scope(value: T, scope: ScopeId) -> CopyValue<T>

Create a new CopyValue. The value will be stored in the given scope. When the specified scope is dropped, the value will be dropped.

impl<T, S> CopyValue<T, S>
where T: 'static, S: Storage<T>,

pub fn new_maybe_sync(value: T) -> CopyValue<T, S>

Create a new CopyValue. The value will be stored in the current component.

Once the component this value is created in is dropped, the value will be dropped.

pub fn leak_with_caller( value: T, caller: &'static Location<'static>, ) -> CopyValue<T, S>

Create a new CopyValue without an owner. This will leak memory if you don’t manually drop it.

pub fn point_to(&self, other: CopyValue<T, S>) -> Result<(), BorrowError>

Point to another copy value

pub fn new_maybe_sync_in_scope(value: T, scope: ScopeId) -> CopyValue<T, S>

Create a new CopyValue. The value will be stored in the given scope. When the specified scope is dropped, the value will be dropped.

pub fn new_maybe_sync_in_scope_with_caller( value: T, scope: ScopeId, caller: &'static Location<'static>, ) -> CopyValue<T, S>

Create a new CopyValue with a custom caller. The value will be stored in the given scope. When the specified scope is dropped, the value will be dropped.

pub fn manually_drop(&self)

Manually drop the value in the CopyValue, invalidating the value in the process.

pub fn origin_scope(&self) -> ScopeId

Get the scope this value was created in.

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the value.

pub fn value(&self) -> GenerationalBox<T, S>

Get the underlying GenerationalBox value.