dioxus::prelude

Struct Global

pub struct Global<T, R = T> { /* private fields */ }

A lazy value that is created once per application and can be accessed from anywhere in that application

Implementations

impl<T> Global<Memo<T>, T>
where T: PartialEq + 'static,

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the signal.

pub fn memo(&self) -> Memo<T>

Resolve the global memo. This will try to get the existing value from the current virtual dom, and if it doesn’t exist, it will create a new one.

impl<T> Global<Signal<T>, T>
where T: 'static,

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the signal.

pub fn signal(&self) -> Signal<T>

Resolve the global signal. This will try to get the existing value from the current virtual dom, and if it doesn’t exist, it will create a new one.

impl<T, R> Global<T, R>
where T: Clone + 'static + Writable<Target = R> + InitializeFromFunction<R>, R: 'static,

pub fn write(&self) -> <T as Writable>::Mut<'static, R>

Write this value

pub fn with_mut<O>(&self, f: impl FnOnce(&mut R) -> O) -> O

Run a closure with a mutable reference to the signal’s value. If the signal has been dropped, this will panic.

impl<T, R> Global<T, R>
where T: Clone + 'static + InitializeFromFunction<R>,

pub const fn new(constructor: fn() -> R) -> Global<T, R>

Create a new global value

pub const fn with_name( constructor: fn() -> R, key: &'static str, ) -> Global<T, R>

Create this global signal with a specific key. This is useful for ensuring that the signal is unique across the application and accessible from outside the application too.

pub const fn with_location( constructor: fn() -> R, file: &'static str, line: u32, column: u32, index: usize, ) -> Global<T, R>

Create this global signal with a specific key. This is useful for ensuring that the signal is unique across the application and accessible from outside the application too.

pub fn key(&self) -> GlobalKey<'static>

Get the key for this global

pub fn resolve(&self) -> T

Resolve the global value. This will try to get the existing value from the current virtual dom, and if it doesn’t exist, it will create a new one.

pub fn origin_scope(&self) -> ScopeId

Get the scope the signal was created in.