dioxus::prelude

Struct LaunchBuilder

pub struct LaunchBuilder { /* private fields */ }

A builder for a fullstack app.

Implementations

impl LaunchBuilder

pub fn new() -> LaunchBuilder

Create a new builder for your application. This will create a launch configuration for the current platform based on the features enabled on the dioxus crate.

pub fn desktop() -> LaunchBuilder

Launch your desktop application.

pub fn custom(launch_fn: LaunchFn) -> LaunchBuilder

Provide a custom launch function for your application.

Useful for third party renderers to tap into the launch builder API without having to reimplement it.

Example

use dioxus::prelude::*;
use std::any::Any;

#[derive(Default)]
struct Config;

fn my_custom_launcher(root: fn() -> Element, contexts: Vec<Box<dyn Fn() -> Box<dyn Any> + Send + Sync>>, cfg: Vec<Box<dyn Any>>) {
    println!("launching with root: {:?}", root());
    loop {
        println!("running...");
    }
}

fn app() -> Element {
    rsx! {
        div { "Hello, world!" }
    }
}

dioxus::LaunchBuilder::custom(my_custom_launcher).launch(app);

impl LaunchBuilder

pub fn with_context_provider( self, state: impl Fn() -> Box<dyn Any> + Send + Sync + 'static, ) -> Self

Inject state into the root component’s context that is created on the thread that the app is launched on.

Example

use dioxus::prelude::*;
use std::any::Any;

#[derive(Default)]
struct MyState {
    value: i32,
}

fn app() -> Element {
    rsx! {
        div { "Hello, world!" }
    }
}

dioxus::LaunchBuilder::new()
    .with_context_provider(|| Box::new(MyState { value: 42 }))
    .launch(app);

pub fn with_context( self, state: impl Any + Clone + Send + Sync + 'static, ) -> Self

Inject state into the root component’s context.

Example

use dioxus::prelude::*;
use std::any::Any;

#[derive(Clone)]
struct MyState {
    value: i32,
}

fn app() -> Element {
    rsx! {
        div { "Hello, world!" }
    }
}

dioxus::LaunchBuilder::new()
    .with_context(MyState { value: 42 })
    .launch(app);

impl LaunchBuilder

pub fn with_cfg(self, config: impl LaunchConfig) -> Self

Provide a platform-specific config to the builder.

Example

use dioxus::prelude::*;
use dioxus_desktop::Config;

fn app() -> Element {
    rsx! {
        div { "Hello, world!" }
    }
}

dioxus::LaunchBuilder::desktop()
    .with_cfg(Config::new().with_window(|w| w.with_title("My App")))
    .launch(app);

pub fn launch(self, app: fn() -> Element)

Launch your application.