dioxus::prelude

Struct PointerData

pub struct PointerData { /* private fields */ }

Implementations

impl PointerData

pub fn new(data: impl HasPointerData + 'static) -> PointerData

Create a new PointerData

pub fn downcast<T>(&self) -> Option<&T>
where T: 'static,

Downcast this event to a concrete event type

impl PointerData

pub fn pointer_id(&self) -> i32

Gets the unique identifier of the pointer causing the event.

pub fn width(&self) -> i32

Gets the width (magnitude on the X axis), in CSS pixels, of the contact geometry of the pointer.

pub fn height(&self) -> i32

Gets the height (magnitude on the Y axis), in CSS pixels, of the contact geometry of the pointer.

pub fn pressure(&self) -> f32

Gets the normalized pressure of the pointer input in the range of 0 to 1,

pub fn tangential_pressure(&self) -> f32

Gets the normalized tangential pressure of the pointer input (also known as barrel pressure or cylinder stress) in the range -1 to 1,

pub fn tilt_x(&self) -> i32

Gets the plane angle (in degrees, in the range of -90 to 90) between the Y-Z plane and the plane containing both the transducer (e.g. pen stylus) axis and the Y axis.

pub fn tilt_y(&self) -> i32

Gets the plane angle (in degrees, in the range of -90 to 90) between the X-Z plane and the plane containing both the transducer (e.g. pen stylus) axis and the X axis.

pub fn twist(&self) -> i32

Gets the clockwise rotation of the pointer (e.g. pen stylus) around its major axis in degrees, with a value in the range 0 to 359.The clockwise rotation of the pointer (e.g. pen stylus) around its major axis in degrees, with a value in the range 0 to 359.

pub fn pointer_type(&self) -> String

Gets the device type that caused the event (mouse, pen, touch, etc.).

pub fn is_primary(&self) -> bool

Gets if the pointer represents the primary pointer of this pointer type.