dioxus::prelude

Struct UseFuture

pub struct UseFuture { /* private fields */ }

Implementations

impl UseFuture

pub fn restart(&mut self)

Restart the future with new dependencies.

Will not cancel the previous future, but will ignore any values that it generates.

pub fn cancel(&mut self)

Forcefully cancel a future

pub fn pause(&mut self)

Pause the future

pub fn resume(&mut self)

Resume the future

pub fn task(&self) -> Task

Get a handle to the inner task backing this future Modify the task through this handle will cause inconsistent state

pub fn finished(&self) -> bool

Is the future currently finished running?

Reading this does not subscribe to the futureโs state

pub fn state(&self) -> ReadOnlySignal<UseFutureState>

Get the current state of the future.