dioxus::prelude

Trait HasResizeData

pub trait HasResizeData: Any {
    // Required method
    fn as_any(&self) -> &(dyn Any + 'static);

    // Provided methods
    fn get_border_box_size(&self) -> Result<Size2D<f64, Pixels>, ResizeError> { ... }
    fn get_content_box_size(&self) -> Result<Size2D<f64, Pixels>, ResizeError> { ... }
}

Required Methods

fn as_any(&self) -> &(dyn Any + 'static)

return self as Any

Provided Methods

fn get_border_box_size(&self) -> Result<Size2D<f64, Pixels>, ResizeError>

Get the border box size of the observed element

fn get_content_box_size(&self) -> Result<Size2D<f64, Pixels>, ResizeError>

Get the content box size of the observed element