dioxus::prelude

Trait ReadableVecExt

pub trait ReadableVecExt<T>: Readable<Target = Vec<T>>
where
    T: 'static,
{
    // Provided methods
    fn len(&self) -> usize { ... }
    fn is_empty(&self) -> bool { ... }
    fn first(&self) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>> { ... }
    fn last(&self) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>> { ... }
    fn get(
        &self,
        index: usize,
    ) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>> { ... }
    fn iter(&self) -> ReadableValueIterator<'_, Self> 
       where Self: Sized { ... }
}

An extension trait for Readable<Vec> that provides some convenience methods.

Provided Methods

fn len(&self) -> usize

Returns the length of the inner vector.

fn is_empty(&self) -> bool

Returns true if the inner vector is empty.

fn first(&self) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>>

Get the first element of the inner vector.

fn last(&self) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>>

Get the last element of the inner vector.

fn get(&self, index: usize) -> Option<<Self::Storage as AnyStorage>::Ref<'_, T>>

Get the element at the given index of the inner vector.

fn iter(&self) -> ReadableValueIterator<'_, Self>
where Self: Sized,

Get an iterator over the values of the inner vector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.