Trait dioxus::prelude::InteractionElementOffset

pub trait InteractionElementOffset: InteractionLocation {
    // Required method
    fn element_coordinates(&self) -> Point2D<f64, ElementSpace>;

    // Provided method
    fn coordinates(&self) -> Coordinates { ... }
}

A interaction that contains data about the location of the event.

Required Methods

fn element_coordinates(&self) -> Point2D<f64, ElementSpace>

Gets the coordinates of the event relative to the target element.

Provided Methods

fn coordinates(&self) -> Coordinates

Gets the coordinates of the event.